Refactor user requests out of user
