+++ lib/alma.rb
@@ -3,6 +3,7 @@ require 'alma/config'
 require 'alma/api'
 require 'alma/error'
 require 'alma/alma_record'
+require 'alma/user_requests'
 require 'alma/user'
 require 'alma/bib'
 require 'alma/loan'
